Make git clone base URL configurable to enable full offline builds.
This is a first stab. I tried using Rebar's deps_dir as per
https://github.com/rebar/rebar/wiki/Dynamic-configuration
but I couldn't get it to recognise paths vs. URLs. Still
there might be a better way.

The eventual goal would be to have a full mirror of all repos
locally and being able to set the dep src for all CouchDB
dependencies via an environment variable. It is simple
enough for the main couchdb repo, but some modules, like
couch_log here define their own dependencies and they should
pick up the local dep src as well.

So yeah, weaving my hands a little, but wanted to throw this
into the ring for discussion. Let me know what you think and
how this can be done better :)

Thanks!